+++ biopipen/scripts/scrna/ExprImputation.R
@@ -1,0 +1,7 @@
+{% if envs.tool == "rmagic" %}
+{% include biopipen_dir + "/scripts/scrna/ExprImputation-rmagic.R" %}
+{% elif envs.tool == "scimpute" %}
+{% include biopipen_dir + "/scripts/scrna/ExprImputation-scimpute.R" %}
+{% elif envs.tool == "alra" %}
+{% include biopipen_dir + "/scripts/scrna/ExprImputation-alra.R" %}
+{% endif %}